requirements.txt
Babel==2.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
pymongo==4.6.0
python-dotenv==1.0.0
setuptools==69.0.2
speaklater==1.3
Werkzeug==3.0.1
WTForms==3.1.1
